Backscatter: scatter % and ionospheric species density have been vetted - they are correct.  Backscatter in test config: kicks out cdf without up/downgoing maxwellian, csv files are saved to disk with intermediate BS steps for diagnostics.  Explicitly converted various int types to avoid compile warnings, added explicit unit typedefs, removed pad from CSV - fileIO handles it now

+++ include/ErrorHandling/SimException.h
@@ -22,7 +22,7 @@ public:
 		if (!numericalArguments_m.empty())
 		{
 			args_m += "double: ";
-			for (unsigned int iii = 0; iii < numericalArguments_m.size(); iii++)
+			for (size_t iii = 0; iii < numericalArguments_m.size(); iii++)
 			{
 				args_m += std::to_string(numericalArguments_m.at(iii));
 				if (iii < numericalArguments_m.size() - 1)
@@ -33,7 +33,7 @@ public:
 		if (!stringArguments_m.empty())
 		{
 			args_m += "; string: ";
-			for (unsigned int iii = 0; iii < stringArguments_m.size(); iii++)
+			for (size_t iii = 0; iii < stringArguments_m.size(); iii++)
 			{
 				args_m += stringArguments_m.at(iii);
 				if (iii < stringArguments_m.size() - 1)
