+++ src/Simulation/previousSimulation.cpp
@@ -59,11 +59,11 @@ Simulation::Simulation(std::string prevSimDir) : saveRootDir_m{ prevSimDir }, /*
 
 
 	//Load EField Model
-	for (unsigned int entry = 0; entry < simAttr_m->EAD.names_m.size(); entry++)
+	for (size_t entry = 0; entry < simAttr_m->EAD.names_m.size(); entry++)
 		addEFieldModel(simAttr_m->EAD.names_m.at(entry), simAttr_m->EAD.dblAttrs_m.at(entry), false);
 
 	//Load Particles
-	for (unsigned int part = 0; part < simAttr_m->partAD.names_m.size(); part++)
+	for (size_t part = 0; part < simAttr_m->partAD.names_m.size(); part++)
 	{
 		createParticleType(
 			simAttr_m->partAD.names_m.at(part),
@@ -76,7 +76,7 @@ Simulation::Simulation(std::string prevSimDir) : saveRootDir_m{ prevSimDir }, /*
 	}
 	
 	//Load Satellites
-	for (unsigned int sat = 0; sat < simAttr_m->satAD.names_m.size(); sat++)
+	for (size_t sat = 0; sat < simAttr_m->satAD.names_m.size(); sat++)
 	{
 		std::unique_ptr<TempSat> tmpsat{ std::make_unique<TempSat>(
 			(int)(simAttr_m->satAD.dblAttrs_m.at(sat).at(findAttrInd("partInd", simAttr_m->satAD.dblLabels_m.at(sat)))),
